Raid and lvm support (#13)
* Remove unfinished code block

* Force cryptsetup in initramfs when LUKS is used

* Remove unused defaults

* Support for DM RAID and LVM, closes #1

* Update documentation for LVM and MD

* Default mkfs options in vars

* Change ssh key variable, correct path

* The variable for the SSH Key presumes the keys have been fetched from
a local SSH agent, that doesn't have to be the case. The new name makes
it clearer for the playbook user.
* The path to the bootstrap_complete indicator file wasn't checked
correctly. This will create the file on the target as well as on the
host where the plays are executed, preventing the role from running
twice.

* Fix location for initramfs-tools config

* Remove Meaningless vars

* USe device name instead of UUID

This seems to fix issues with device detection when nesting DM-RAID,
LUKS and LVM

* LVM / MD compatibility

* Make device layering work
* groundwork for supporting different filesystems
* remove specific tasks that are more suited for a playbook (grub)
* make bootstrap environment user accesible to facilitate running roles
* Cleaning up the cleanup
* initial work on namespacing