Version 4 (#18)
* Remove unfinished code block

* Force cryptsetup in initramfs when LUKS is used

* Remove unused defaults

* Support for DM RAID and LVM, closes #1

* Update documentation for LVM and MD

* Default mkfs options in vars

* Change ssh key variable, correct path

* The variable for the SSH Key presumes the keys have been fetched from
a local SSH agent, that doesn't have to be the case. The new name makes
it clearer for the playbook user.
* The path to the bootstrap_complete indicator file wasn't checked
correctly. This will create the file on the target as well as on the
host where the plays are executed, preventing the role from running
twice.

* Fix location for initramfs-tools config

* Remove Meaningless vars

* USe device name instead of UUID

This seems to fix issues with device detection when nesting DM-RAID,
LUKS and LVM

* LVM / MD compatibility

* Make device layering work
* groundwork for supporting different filesystems
* remove specific tasks that are more suited for a playbook (grub)
* make bootstrap environment user accesible to facilitate running roles
* Cleaning up the cleanup
* initial work on namespacing

* Add sudo to requried packages

Without sudo the SSH chrooting won't work.

* Skip devices without a logical volume

If you create a thin pool or a cache device the device may not have a
volume name and as such isn't mounted. These devices can be skipped.

* Correct LUKS1 version check

The LUKS1 task was always run, overwriting the LUKS2 device that was
previously created. This corrects the version check.

* Change task name to reflect all filesystems

* Allow discards on LUKS device

* Support for NVMe devices

Since the enumeration of NVMe devices works differently the role had
issues detecting the corresponding device name for partitions created on
NVMe devices. This should fix those path names.

* Support for apt proxy

Adds support for an apt proxy running on the vagrant host, an nginx
configuration file will be provided elsewhere.

* newline at eof

* Remove syntax error in cleanup

* Prevent leakage of passphrase

Switches over to the command module for tasks dealing with the password
and also disables log output so the passphrase isn't leaked by accident.

* Support root on thinly provisioned volume

This allows using a thinly provisioned LVM volume as root volume,
requires some initramfs adjustments. The test inventory has been
adjusted so this can easily be tested.

* Remove obsolete Limitation from README

* Add sudo to requried packages

Without sudo the SSH chrooting won't work.

* Skip devices without a logical volume

If you create a thin pool or a cache device the device may not have a
volume name and as such isn't mounted. These devices can be skipped.

* Correct LUKS1 version check

The LUKS1 task was always run, overwriting the LUKS2 device that was
previously created. This corrects the version check.

* Change task name to reflect all filesystems

* Allow discards on LUKS device

* Support for NVMe devices

Since the enumeration of NVMe devices works differently the role had
issues detecting the corresponding device name for partitions created on
NVMe devices. This should fix those path names.

* Support for apt proxy

Adds support for an apt proxy running on the vagrant host, an nginx
configuration file will be provided elsewhere.

* newline at eof

* Remove syntax error in cleanup

* Prevent leakage of passphrase

Switches over to the command module for tasks dealing with the password
and also disables log output so the passphrase isn't leaked by accident.

* Support root on thinly provisioned volume

This allows using a thinly provisioned LVM volume as root volume,
requires some initramfs adjustments. The test inventory has been
adjusted so this can easily be tested.

* * Remove documentation for obsolete parameter
* Clean up spelling for use_efi options

* Disable EFI when no partition is available

* Add python3-minimal to packages

This makes it possible to run ansible on python3 as well.

* Fix EFI detection

* Fix LVM cleanup

Deactivating LVM devices should be performed in reverse order to
make sure thin pools are removed after devices using the thin
pool.

* Fix proxy cache detection

* Fix support for zfs 0.8

* Try not to output password on console

* Include hostid from Host

Without having the hostid set the import at boot will fail

* Debootstrap user improvements (#21)

* Improved of creating debootstrap user
* Added notes related to user options